requirements.txt
alembic==1.13.1
click==8.1.7
joblib==1.4.0
Mako==1.3.3
MarkupSafe==2.1.5
nltk==3.8.1
psycopg2==2.9.9
python-dotenv==1.0.1
regex==2024.4.28
SQLAlchemy==2.0.29
tqdm==4.66.2
typing_extensions==4.11.0
